+++ src/sad_spirit/pg_wrapper/converters/BaseConverter.php
@@ -14,10 +14,7 @@ declare(strict_types=1);
 
 namespace sad_spirit\pg_wrapper\converters;
 
-use sad_spirit\pg_wrapper\{
-    TypeConverter,
-    exceptions\TypeConversionException
-};
+use sad_spirit\pg_wrapper\TypeConverter;
 
 /**
  * Base class for type converters, defines handling for null values and some parsing helpers
@@ -54,33 +51,4 @@ abstract class BaseConverter implements TypeConverter
      * Converts PHP variable not identical to null into native format
      */
     abstract protected function outputNotNull(mixed $value): string;
-
-    /**
-     * Gets next non-whitespace character from input
-     *
-     * @param string $str Input string
-     * @param int    $p   Position within input string
-     * @return string|null
-     */
-    protected function nextChar(string $str, int &$p): ?string
-    {
-        $p += \strspn($str, self::WHITESPACE, $p);
-        return $str[$p] ?? null;
-    }
-
-    /**
-     * Throws an Exception if next non-whitespace character in input is not the given char
-     *
-     * @param string $string
-     * @param int    $pos
-     * @param string $char
-     * @throws TypeConversionException
-     */
-    protected function expectChar(string $string, int &$pos, string $char): void
-    {
-        if ($char !== $this->nextChar($string, $pos)) {
-            throw TypeConversionException::parsingFailed($this, "'" . $char . "'", $string, $pos);
-        }
-        $pos++;
-    }
 }