+++ src/sad_spirit/pg_wrapper/converters/ContainerConverter.php
@@ -21,6 +21,35 @@ use sad_spirit\pg_wrapper\exceptions\TypeConversionException;
  */
 abstract class ContainerConverter extends BaseConverter
 {
+    /**
+     * Gets next non-whitespace character from input
+     *
+     * @param string $str Input string
+     * @param int    $p   Position within input string
+     * @return string|null
+     */
+    protected function nextChar(string $str, int &$p): ?string
+    {
+        $p += \strspn($str, self::WHITESPACE, $p);
+        return $str[$p] ?? null;
+    }
+
+    /**
+     * Throws an Exception if next non-whitespace character in input is not the given char
+     *
+     * @param string $string
+     * @param int    $pos
+     * @param string $char
+     * @throws TypeConversionException
+     */
+    protected function expectChar(string $string, int &$pos, string $char): void
+    {
+        if ($char !== $this->nextChar($string, $pos)) {
+            throw TypeConversionException::parsingFailed($this, "'" . $char . "'", $string, $pos);
+        }
+        $pos++;
+    }
+
     protected function inputNotNull(string $native): mixed
     {
         $pos   = 0;